+++ opentelemetry-sdk/CHANGELOG.md
@@ -18,6 +18,8 @@
   ([#1237](https://github.com/open-telemetry/opentelemetry-python/pull/1237))
 - Allow samplers to modify tracestate
   ([#1319](https://github.com/open-telemetry/opentelemetry-python/pull/1319))
+- Remove TracerProvider coupling from Tracer init
+  ([#1295](https://github.com/open-telemetry/opentelemetry-python/pull/1295))
 
 ## Version 0.14b0